CORE-2731: split the color audit engine from its REX-specific half
src/test/cssColors.ts is now pure CSS parsing with no imports at all: given
stylesheet text, which color literals are in it and which declaration each was
written in. Everything that knows about REX -- the theme token index, the
off-palette allowlist, the stylesheet walk and the violation report -- moves to
src/test/themeColors.ts, which is what theme.spec.ts and the baseline generator
use.

The boundary already existed; only the file did not. fs and path were used by
the REX half alone, and the engine's half of the spec never needed the theme.
Splitting it lets the parser be reviewed on its own, which is what Roy asked
for on #3133.

No behaviour change: theme.baseline.json regenerates byte-identical.
